JSON config file for SendEmail Obj
