+++ Flask/email_functionality/SendEmailObjAttributes.json
@@ -1,0 +1,8 @@
+{
+  "fromAddr": "yourEmail@gmail.com",
+  "toAddr": "yourEmail@gmail.com",
+  "subject": "Analysed Email Headers Attached.",
+  "text" : "Thank you for using the Eml Parse Py Email Analysis Application.",
+  "attachment": "test.html",
+  "passwd": "password"
+}
